images: tumbleweed: Fix dbus sessions on ssh login
Cloud-init and tumbleweed's setup script created sshd_config file to
path that's not standard on tumbleweed. The creation of sshd_config to
the wrong location accidentally disabled dbus session on ssh login which
caused non-root users to not be able to use systemctl --user command.

On tumbleweed sshd_config is located at /usr/etc/ssh/ instead of
/etc/ssh/. Creating the config file at /etc/ssh/ will override the
default configuration.